12-spotify-playlist/requirements.txt
beautifulsoup4
requests
python-dotenv
spotipy